# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent JSON Reliability",
    "description": "Agent JSON Reliability. Free: inspect, validate, repair, MCP. Paid hosted reliability endpoint: POST /v1/json/reliable (0.003 USDC per call via x402 on Base). Inspect, validate, repair, and MCP are not paid x402 endpoints.",
    "version": "0.1.0",
    "x-llm-required": false,
    "x-guidance": "Free routes: GET /health, POST /v1/json/inspect, POST /v1/json/validate, POST /v1/json/repair, POST /mcp. Paid route: POST /v1/json/reliable. Unpaid calls return HTTP 402. Body: {\"text\": \"<json or malformed json>\", \"schema\": {optional JSON Schema}}."
  },
  "paths": {
...
    "/v1/json/reliable": {
      "post": {
        "summary": "Paid x402 JSON reliability",
        "operationId": "reliable_v1_json_reliable_post",
        "responses": {
          "200": {
            "description": "JSON reliability result after payment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid_original": {
                      "type": "boolean"
                    },
                    "repaired": {
                      "type": "boolean"
                    },
                    "valid_final": {
                      "type": "boolean"
                    },
                    "schema_valid": {
                      "type": "boolean"
                    },
                    "unsafe_or_ambiguous": {
                      "type": "boolean"
                    },
                    "json": {},
                    "errors": {
                      "type": "array"
                    },
                    "changes": {
                      "type": "array"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        },
        "security": [],
        "description": "Paid x402 v2 exact route. Unpaid requests return HTTP 402 with PAYMENT-REQUIRED. 0.003 USDC on Base (eip155:8453). Inspect, validate, repair, and MCP remain free.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.003"
          },
          "protocols": [
            {
              "x402": {
                "scheme": "exact",
                "network": "eip155:8453",
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "amount": "3000",
                "payTo": ""
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "Malformed or valid JSON text from an agent"
                  },
                  "schema": {
                    "type": "object",
                    "description": "Optional JSON Schema"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8770",
      "description": "Public origin"
    }
  ],
  "security": [],
...
}